refactor(api): Split up utils.js and document the utils
